requirements.txt
#pymongo
#dnspython
pymongo[srv]
python-dotenv
xmltodict
requests
tabulate